Input parser logic.
* Rename of input_parser to command_line_parser.
* Finish command_line_parser.
* Document missing methods.
* Rename init class.
