+++ init.py
@@ -1,8 +1,10 @@
 from input_parser.input_parser_factory import InputParserFactory
 
-class Main():
+class InitializerWrapper():
   def __init__(self):
-    InputParserFactory.get_parser()
+    parser = InputParserFactory.get_parser()
+    arguments = parser.get_input_params()
+    print(arguments)
 
 if __name__ == '__main__':
-  main = Main()
+  initializer = InitializerWrapper()
