+++ input_parser/iinput_parser.py
@@ -4,4 +4,5 @@ class IInputParser(metaclass=abc.ABCMeta):
 
   @abc.abstractmethod
   def get_input_params(self) -> dict:
+    """Returns the list of arguments used to run the script."""
     pass
