+++ input_parser/input_parser_factory.py
@@ -1,6 +1,11 @@
-from input_parser.input_parser import InputParser
+from input_parser.command_line_parser import CommandLineParser
 
 class InputParserFactory():
+  """Is in charge to create the proper input parser
+  
+  TODO: Implement the factory properly following the pattern.
+  """
   @staticmethod
   def get_parser():
-    return InputParser()
+    """Returns an instance of the proper input parser."""
+    return CommandLineParser()
